+++ src/lib.rs
@@ -65,6 +65,9 @@ pub use crate::algorithms::operators::intersection::intersection;
 pub use crate::algorithms::operators::simplify::simplify;
 pub use crate::algorithms::operators::union::union;
 pub use crate::algorithms::paths::astar::a_star_path;
+pub use crate::algorithms::paths::bellman_ford::{
+    bellman_ford_distances, bellman_ford_distances_with_mode,
+};
 pub use crate::algorithms::paths::dijkstra::{
     DijkstraAllPaths, DijkstraMode, DijkstraPaths, dijkstra_all_shortest_paths, dijkstra_distances,
     dijkstra_distances_cutoff, dijkstra_distances_cutoff_with_mode, dijkstra_distances_multi,